Rename dfrefs checks
